requirements.txt
Django==5.2.8
djangorestframework==3.16.1
gunicorn==23.0.0
whitenoise==6.11.0

openai==2.11.0
openai-whisper==20250625

psycopg[binary]==3.3.2
python-dotenv==1.2.1
dj-database-url==2.3.0

channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
redis==5.0.8